+++ HomeProject/FoodOrderingSolution/PublicApi.DTO.v1/DrinkDTOs/DrinkCreateDTO.cs
@@ -7,7 +7,7 @@ namespace PublicApi.DTO.v1.DrinkDTOs
     {
         public Guid Id { get; set; }
         
-        public int Size { get; set; } = default!;
+        public float Size { get; set; } = default!;
 
         [MaxLength(256)] [MinLength(1)] public string Name { get; set; } = default!;
